Move doc tests to `nif` sub project
Remove predefined `documented` sub project
