--- test_projects/documented/c_src/example/_generated/.gitignore
@@ -1,3 +1,0 @@
-**/*.h
-**/*.c
-**/*.cpp
