--- test_projects/documented/c_src/example/example.h
@@ -1,1 +1,0 @@
-#include "_generated/example.h"
