--- test_projects/documented/lib/example.ex
@@ -1,3 +1,0 @@
-defmodule Example do
-  use Unifex.Loader
-end
